+++ commands/transfer_funds.go
@@ -292,4 +292,9 @@ func validateDispatchStrategy(toAccountType vega.AccountType, dispatchStrategy *
 			}
 		}
 	}
+
+	// temporarily disable teams scope
+	if dispatchStrategy.EntityScope == vega.EntityScope_ENTITY_SCOPE_TEAMS {
+		errs.AddForProperty(prefix+".team_scope", ErrIsNotSupported)
+	}
 }